Resolve bugs with module loading
Alters the module cache to use an MD5 hash of the loaded module addresses instead of the number of modules. This resolves an issue where a process may unload and load a different module such that the module count (and resulting hash) remains the same.

Upon an error with loading the modues, we don't need to fully crash out of the function. The exception can be handled gracefully and the rest of the modules loaded. On the next invocation the handles will hopefully be sorted out without problems.